Added Part of Basic UI system. Ability to manipulate block qualities, add, and remove blocks without going into code.

+++ include/physics.h
@@ -7,8 +7,8 @@ class Physics {
 public:
 	void DeepSpaceHeatTransfer(std::vector <ThermalBlocks > * block, float dt, float transferspeed, double sigma, double emissivety);
 	//float HeatTransfer();
-	void AddConduction(const ThermalBlocks& BlockA, const ThermalBlocks& BlockB, size_t& ba, size_t& bb, float dt, float transferspeed, std::vector<float>& tempsToadd, BlockManager& blockUser);
-	void AddRadiation(const ThermalBlocks& BlockA, const ThermalBlocks& BlockB, size_t& ba, size_t& bb, float dt, float transferspeed, float sigma, std::vector<float>& tempsToadd, BlockManager& blockUser);
+	void AddConduction(const ThermalBlocks& BlockA, const ThermalBlocks& BlockB, size_t& ba, size_t& bb, float dt, float transferspeed, std::vector<double>& tempsToadd, BlockManager& blockUser);
+	void AddRadiation(const ThermalBlocks& BlockA, const ThermalBlocks& BlockB, size_t& ba, size_t& bb, float dt, float transferspeed, double sigma, std::vector<double>& tempsToadd, BlockManager& blockUser);
 	void GetRGB(ThermalBlocks* blocks);;
 	void getCoolingConstant(double sigma, std::vector<ThermalBlocks>& blocksinuse);
 };
